+++ components/molecules/CardRepoList/card-repo-list.tsx
@@ -45,7 +45,7 @@ const CardRepoList = ({
                       <div className="flex gap-1  p-1 pr-2 border-[1px] border-light-slate-6 rounded-lg text-light-slate-12">
                         <Icon IconImage={repoIcon} className="rounded-[4px] overflow-hidden" />
                         <span className={`max-w-[45px] md:max-w-[100px] truncate ${fontSizeClassName}`}>
-                          <Link href={`https://github.com/${repoOwner}/${repoName}`} target="_blank" rel="noreferrer">
+                          <Link href={`/s/${repoOwner}/${repoName}`} target="_blank">
                             {repoName}
                           </Link>
                         </span>