+++ components/molecules/RecommendedRepoCard/recommended-repo-card.tsx
@@ -33,12 +33,7 @@ const RecommendedRepoCard = ({ fullName, className }: RecommendedRepoCardProps):
       {data && (
         <>
           <div className="flex justify-between w-full">
-            <a
-              target="_blank"
-              href={`https://github.com/${fullName}`}
-              className="flex items-center gap-1.5"
-              rel="noreferrer"
-            >
+            <Link href={`/s/${fullName}`} className="flex items-center gap-1.5">
               <picture>
                 <img
                   alt="Hot Repo Icon"
@@ -47,7 +42,7 @@ const RecommendedRepoCard = ({ fullName, className }: RecommendedRepoCardProps):
                 />
               </picture>
               <span className="text-sm text-light-slate-11">{owner}</span>
-            </a>
+            </Link>
           </div>
 
           <div className="flex flex-col gap-2 pb-3">