+++ next.config.js
@@ -135,7 +135,7 @@ module.exports = withSentryConfig(
     silent: true,
 
     org: "opensauced",
-    project: "insights",
+    project: process.env.NODE_ENV === "production" ? "app-opensauced" : "beta-app-opensauced",
   },
   {
     // For all available options, see:
@@ -145,7 +145,7 @@ module.exports = withSentryConfig(
     widenClientFileUpload: true,
 
     // Transpiles SDK to be compatible with IE11 (increases bundle size)
-    transpileClientSDK: true,
+    transpileClientSDK: false,
 
     // Routes browser requests to Sentry through a Next.js rewrite to circumvent ad-blockers (increases server load)
     tunnelRoute: "/monitoring",