+++ pages/s/[org]/[repo]/index.tsx
@@ -1,6 +1,7 @@
 import { GetServerSidePropsContext } from "next";
 import { useRouter } from "next/router";
 import { ComponentProps } from "react";
+import { HiOutlineExternalLink } from "react-icons/hi";
 import { fetchApiData } from "helpers/fetchApiData";
 import { useFetchMetricStats } from "lib/hooks/api/useFetchMetricStats";
 
@@ -24,7 +25,7 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     return { notFound: true };
   }
 
-  const response = await fetch(repoData.url);
+  const response = await fetch(repoData.url || `https://api.github.com/repos/${org}/${repo}`);
   const { owner } = await response.json();
 
   const range = (context.query.range ? Number(context.query.range) : 30) as Range;
@@ -79,7 +80,14 @@ export default function RepoPage({ repoData, image, ogImageUrl }: RepoPageProps)
           <header className="flex items-center gap-4">
             <Avatar size={96} avatarURL={image} />
             <div className="flex flex-col gap-2">
-              <h1 className="text-xl md:text-3xl font-bold">{repoData.full_name}</h1>
+              <a
+                href={`https://github.com/${repoData.full_name}`}
+                target="_blank"
+                className="group hover:underline underline-offset-2 text-xl md:text-3xl font-bold flex gap-2 items-center"
+              >
+                <h1>{repoData.full_name}</h1>
+                <HiOutlineExternalLink className="group-hover:text-sauced-orange text-lg lg:text-xl" />
+              </a>
               <p className="md:text-xl">{repoData.description}</p>
             </div>
           </header>