+++ sentry.edge.config.ts
@@ -6,10 +6,12 @@
 import * as Sentry from "@sentry/nextjs";
 
 Sentry.init({
+  dsn: process.env.SENTRY_DSN || process.env.NEXT_PUBLIC_SENTRY_DSN,
+
   enabled: process.env.NODE_ENV !== "development",
 
   // Adjust this value in production, or use tracesSampler for greater control
-  tracesSampleRate: 0.1,
+  tracesSampleRate: 1,
 
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,